requirements.txt
alembic==1.7.7
click==8.0.4
dnspython==2.2.1
email-validator==1.1.3
Flask==2.0.3
Flask-Login==0.6.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.0
itsdangerous==2.1.1
Jinja2==3.0.3
Mako==1.2.0
MarkupSafe==2.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-dotenv==0.19.2
SQLAlchemy==1.4.35
Werkzeug==2.0.3
WTForms==3.0.1
zipp==3.8.0